requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
openai>=1.30.0
faiss-cpu>=1.8.0
streamlit>=1.35.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.13.0
plotly>=5.21.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
joblib>=1.4.0
